integrate admin dashboard
